* commit '10ba4891daa857ef58d57ae6df86bb3475b82202':
  DM-16058: Optimize BDATA Search query
  DM-16058: Optimize BDATA Search query
  DM-16058: Optimize BDATA Search query
  DM-16058: Optimize BDATA Search query
  DM-16058: Optimize BDATA Search query
  Added initial version of the herd.postgres.0.175.0-to-0.176.0.upgrade.sql script
  CATDM-1675: Change Storage Unit ID field from int to long
  Update Herd pom version to 176
  CATDM-1675: Change Storage Unit ID field from int to long
  CATDM-1675: Change Storage Unit ID field from int to long

+++ herd-code/herd-dao/src/main/java/org/finra/herd/dao/StorageFileDao.java
@@ -51,7 +51,7 @@ public interface StorageFileDao extends BaseJpaDao
      *
      * @return the map of storage unit ids to their corresponding storage file paths.
      */
-    MultiValuedMap<Integer, String> getStorageFilePathsByStorageUnitIds(List<Integer> storageUnitIds);
+    MultiValuedMap<Long, String> getStorageFilePathsByStorageUnitIds(List<Long> storageUnitIds);
 
     /**
      * Saves a list of storage file entities.
